Split compio-driver out.
