+++ compio-driver/Cargo.toml
@@ -1,0 +1,45 @@
+[package]
+name = "compio-driver"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+# Workspace dependencies
+compio-buf = { workspace = true }
+
+# Utils
+cfg-if = "1"
+# may be excluded from linking if the unstable equivalent is used
+once_cell = "1"
+slab = "0.4"
+socket2 = { version = "0.5", features = ["all"] }
+
+# Windows specific dependencies
+[target.'cfg(target_os = "windows")'.dependencies]
+compio-buf = { workspace = true, features = ["arrayvec"] }
+widestring = "1"
+windows-sys = { version = "0.48", features = [
+    "Win32_Foundation",
+    "Win32_Networking_WinSock",
+    "Win32_Security",
+    "Win32_Storage_FileSystem",
+    "Win32_System_Console",
+    "Win32_System_IO",
+    "Win32_System_Pipes",
+    "Win32_System_SystemServices",
+    "Win32_System_Threading",
+] }
+
+# Linux specific dependencies
+[target.'cfg(target_os = "linux")'.dependencies]
+io-uring = { version = "0.6", optional = true }
+polling = { version = "3", optional = true }
+libc = "0.2"
+
+# Other platform dependencies
+[target.'cfg(all(not(target_os = "linux"), unix))'.dependencies]
+polling = "3"
+libc = "0.2"
+
+[features]
+default = ["io-uring"]
